+++ packages/bluesky/lib/src/moderation/definitions/label_groups.dart
@@ -25,6 +25,7 @@ const labelGroupDefinitionSystem = LabelGroupDefinition(
     labelDefinitionHide,
     labelDefinitionNoPromote,
     labelDefinitionWarn,
+    labelDefinitionNoUnauthenticated,
   ],
 );