+++ packages/bluesky/lib/src/moderation/entities/label_definition.g.dart
@@ -57,6 +57,7 @@ const _$LabelPreferenceEnumMap = {
 const _$LabelDefinitionFlagEnumMap = {
   LabelDefinitionFlag.noOverride: 'no-override',
   LabelDefinitionFlag.adult: 'adult',
+  LabelDefinitionFlag.unauthed: 'unauthed',
 };
 
 const _$LabelDefinitionOnWarnBehaviorEnumMap = {